pyproject.toml
# Poetry pyproject.toml: https://python-poetry.org/docs/pyproject/

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "scroll-shelter"
version = "0.1.0"
description = "Awesome `scroll-shelter` is a Python cli/package created with https://github.com/TezRomacH/python-package-template"
readme = "README.md"
authors = [
  "scroll-shelter < >"
]
license = "GNU GPL v3.0"
repository = "https://github.com/jb-delafosse/scroll-shelter"
homepage = "https://github.com/jb-delafosse/scroll-shelter"

# Keywords description https://python-poetry.org/docs/pyproject/#keywords
keywords = []  # Update me

# Pypi classifiers: https://pypi.org/classifiers/
classifiers = [  # Update me
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Operating System :: OS Independent",
  "Topic :: Software Development :: Libraries :: Python Modules",
]

[tool.poetry.scripts]
# Entry points for the package https://python-poetry.org/docs/pyproject/#scripts
"scroll-shelter" = "app.entrypoints.http_server:APP"

[tool.poetry.dependencies]
python = "^3.8"
importlib_metadata = {version = "^1.6.0", python = "<3.8"}
typer = {extras = ["all"], version = "^0.3.2"}
rich = "^9.8.2"
google-api-python-client = "^2.3.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^0.4.4"
fastapi-users = {extras = ["mongodb"], version = "^6.0.0"}
httpx-oauth = "^0.3.5"
uvicorn = {extras = ["standard"], version = "^0.13.4"}
fastapi = "^0.63.0"

[tool.poetry.dev-dependencies]
darglint = "^1.5.8"
isort = "^5.7.0"
pyupgrade = "^2.7.4"
black = "^20.8b1"
mypy = "^0.812"
bandit = "^1.7.0"
safety = "^1.10.3"
pytest = "^6.2.1"
pylint = "^2.6.0"
pydocstyle = "^5.1.1"
pre-commit = "^2.9.3"

[tool.black]
# https://github.com/psf/black
line-length = 80
target-version = ["py38"]

[tool.isort]
# https://github.com/timothycrosley/isort/
known_typing = "typing,types,typing_extensions,mypy,mypy_extensions"
sections = "FUTURE,TYPING,STDLIB,THIRDPARTY,FIRSTPARTY,LOCALFOLDER"
include_trailing_comma = true
default_section = "FIRSTPARTY"
multi_line_output = 3
indent = 4
force_grid_wrap = 0
use_parentheses = true
line_length = 80
